Fixed a bug in skeleton extension where wrong extra length was being measured. Still seeing some extensions fail for mysterious reasons - check D20200416T132058.245055.
We now correct eye masks as well, although I've seen one where there still appear to be holes in the mask. Also throw away overlapping eye masks - we could probably choose the largest to keep but we have so many eyes measured I think we can afford to lose a few dubious ones.

We now build up lists of which body parts belong to which bodies, and remove parts that don't have a body and bodies with no parts. Still need to figure out what to do when a body has *too many* parts.