Added support for bayer images with endings other than '.silc'. Performing background calculation and RoI extraction can now be done from GUI. Background calculation doesn't use a nice moving stack like pysilcam at the moment, but calculates stack for each frame (I think). This is slow.
Need to add support for different image size - added a function to change canvas size but not sure where to call it yet.

Threw together a utility script for viewing images from Bernard and writing out ones that seem to contain particles, but it should probably be superseded by the changes I'm making now.