Added a pretty rough and ready ellipse drawing tool to help with egg annotation.
Right now clicking while the tool is active will erase the existing layer and draw an ellipse over the top, slightly tricky to position. Should ideally not destroy the existing layer and have guides extending from the cursor to the image edges in x and y to show you where to position the top-left point.